Move <NotificationPermissionWarning> into PrivateRoute.tsx
